+++ test/components/accordion/accordion.puppeteer-spec.js
@@ -42,7 +42,7 @@ describe('Accordion Puppeteer Test', () => {
 
     it('should have ajax data in headers', async () => {
       await page.click('#ajax-accordion .accordion-header button');
-
+      await page.waitForTimeout(100);
       page.waitForSelector('#ajax-accordion .accordion-pane.is-expanded > .accordion-header:first-child', { visible: true });
       await page.waitForTimeout(600);